+++ src/widgets/splits/Split.cpp
@@ -875,18 +875,7 @@ void Split::explainSplitting()
 
 void Split::popup()
 {
-    auto app = getApp();
-    Window &window = app->windows->createWindow(WindowType::Popup);
-
-    Split *split = new Split(static_cast<SplitContainer *>(
-        window.getNotebook().getOrAddSelectedPage()));
-
-    split->setChannel(this->getIndirectChannel());
-    split->setModerationMode(this->getModerationMode());
-    split->setFilters(this->getFilters());
-
-    window.getNotebook().getOrAddSelectedPage()->appendSplit(split);
-    window.show();
+    getApp()->windows->createChannelWindow(this);
 }
 
 void Split::clear()