+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 - Minor: Sorted usernames in /vips message to be case-insensitive. (#3696)
 - Minor: Strip leading @ and trailing , from usernames in the `/block` and `/unblock` commands. (#3816)
 - Minor: Added option to open a user's chat in a new tab from the usercard profile picture context menu. (#3625)
+- Minor: Added option to open a user's chat in a new tab/split/window by CTRL/ALT/SHIFT+Left Clicking (#3766)
 - Minor: Fixed tag parsing for consecutive escaped characters. (#3711)
 - Minor: Prevent user from entering incorrect characters in Live Notifications channels list. (#3715, #3730)
 - Minor: Fixed automod caught message notice appearing twice for mods. (#3717)